add little changes in layout and add navigation in travels history for the user data and add navigation to the chat screen
