--- vooazProjeto/app/src/main/java/com/vooazdomain/Vooaz/telas/azchat/azChat.kt
@@ -1,2 +1,0 @@
-package com.vooazdomain.Vooaz.telas.azchat
-
