+++ vooazProjeto/app/src/main/java/com/vooazdomain/Vooaz/telas/destinationsScreen/infoDestination/RouteTravel.kt
@@ -1,6 +1,5 @@
-package com.vooazdomain.Vooaz.telas.destinationsScreen
+package com.vooazdomain.Vooaz.telas.destinationsScreen.infoDestination
 
-import DestinationDetailsScreen
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -13,7 +12,6 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.LocationOn
 import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
@@ -25,7 +23,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.vooazdomain.Vooaz.R
